Add files via upload
we introduce a possible approach to the Audio Intent Classification problem. In particular, the proposed consists in extracting the MFCCs feature of each audio signal and balance the data and used as input for two classification models. The solution compares these two different classification models and chose the best model.